[IFRT] Use AttributeMap wherever applicable in IFRT
This change uses `xla::ifrt::AttributeMap` in IFRT `Client`, `Device`, and
`Topology` attributes. It adds utility functions to convert from/to a map of
`PjRtValueType` that is used in PjRt and was used in these IFRT classes.

IFRT Proxy also uses a native serialization of `xla::ifrt::AttributeMap` in
Version 4 with an added compatibility for previously supported protocol
versions (the minimum version does not change).

PiperOrigin-RevId: 644184301